move type min/max functions to builtins.py
