+++ numba/ir_utils.py
@@ -1240,79 +1240,6 @@ def find_const(func_ir, var):
     require(isinstance(var_def, ir.Const))
     return var_def.value
 
-def get_type_max_value(typ):
-    if isinstance(typ, types.Float):
-        bw = typ.bitwidth
-        if bw == 32:
-            return numpy.finfo(numpy.float32).max
-        if bw == 64:
-            return numpy.finfo(numpy.float64).max
-        raise NotImplementedError("Unsupported floating point type")
-    if isinstance(typ, types.Integer):
-        return typ.maxval
-    raise NotImplementedError("Unsupported type")
-
-def get_type_min_value(typ):
-    if isinstance(typ, types.Float):
-        bw = typ.bitwidth
-        if bw == 32:
-            return numpy.finfo(numpy.float32).min
-        if bw == 64:
-            return numpy.finfo(numpy.float64).min
-        raise NotImplementedError("Unsupported floating point type")
-    if isinstance(typ, types.Integer):
-        return typ.minval
-    raise NotImplementedError("Unsupported type")
-
-@infer_global(get_type_min_value)
-@infer_global(get_type_max_value)
-class MinValInfer(AbstractTemplate):
-    def generic(self, args, kws):
-        assert not kws
-        assert len(args) == 1
-        assert isinstance(args[0], types.DType)
-        return signature(args[0].dtype, *args)
-
-@lower_builtin(get_type_min_value, types.DType)
-def lower_get_type_min_value(context, builder, sig, args):
-    typ = sig.args[0].dtype
-    bw = typ.bitwidth
-
-    if isinstance(typ, types.Integer):
-        lty = lir.IntType(bw)
-        val = typ.minval
-        res = lir.Constant(lty, val)
-    elif isinstance(typ, types.Float):
-        if bw == 32:
-            lty = lir.FloatType()
-        elif bw == 64:
-            lty = lir.DoubleType()
-        else:
-            raise NotImplementedError("llvmlite only supports 32 and 64 bit floats")
-        npty = getattr(numpy, 'float{}'.format(bw))
-        res = lir.Constant(lty, numpy.finfo(npty).min)
-    return impl_ret_untracked(context, builder, lty, res)
-
-@lower_builtin(get_type_max_value, types.DType)
-def lower_get_type_min_value(context, builder, sig, args):
-    typ = sig.args[0].dtype
-    bw = typ.bitwidth
-
-    if isinstance(typ, types.Integer):
-        lty = lir.IntType(bw)
-        val = typ.maxval
-        res = lir.Constant(lty, val)
-    elif isinstance(typ, types.Float):
-        if bw == 32:
-            lty = lir.FloatType()
-        elif bw == 64:
-            lty = lir.DoubleType()
-        else:
-            raise NotImplementedError("llvmlite only supports 32 and 64 bit floats")
-        npty = getattr(numpy, 'float{}'.format(bw))
-        res = lir.Constant(lty, numpy.finfo(npty).max)
-    return impl_ret_untracked(context, builder, lty, res)
-
 def compile_to_numba_ir(mk_func, glbls, typingctx=None, arg_typs=None,
                         typemap=None, calltypes=None):
     """
